refactor(issue69): move pipeline_batch implementation under src/pipeline
